refactor: rename vllm_completion to chat_completion and extract reasoning functionality into a new module
