+++ aidial_adapter_openai/chat_completions/vllm/__init__.py
@@ -1,2 +1,3 @@
-from .vllm_completion import chat_completion, extract_reasoning
+from .chat_completion import chat_completion
+from .extract_reasoning import extract_reasoning
 from .vllm_tokenizer import VllmTokenizer
